- Changes to MFIni, MFString - Added MFList<>
git-svn-id: https://fuji.googlecode.com/svn/trunk@670 7e75b669-355e-e21c-2746-f45317fb21cf

+++ Fuji/Private/Include/MFPtrList.inl
@@ -286,7 +286,7 @@ inline int MFPtrListDL<T>::GetLength()
 {
 	T** iterator = ppMark;
 	while(*iterator!=0) ++iterator;
-	return iterator - ppMark;
+	return (int)(iterator - ppMark);
 }
 
 template<class T>
